# Add Buffered BinaryReader to speed up Binlog reading