--- a/src/Framework/Logging/IBinaryReader.cs
+++ b/src/Framework/Logging/IBinaryReader.cs
@@ -22,6 +22,8 @@ namespace Microsoft.Build.Framework.Logging
 
         public byte[] ReadBytes(int count);
 
+        public byte[] ReadGuid();
+
         public bool ReadBoolean();
 
         public long ReadInt64();