--- a/src/Shared/BinaryWriterExtensions.cs
+++ b/src/Shared/BinaryWriterExtensions.cs
@@ -86,7 +86,6 @@ namespace Microsoft.Build.Shared
             }
 
             writer.Write((byte)v);
-            return;
         }
 
 #if !TASKHOST