… src/Build/Logging/BinaryLogger/BuildEventArgsWriter.cs
… src/Build/Logging/BinaryLogger/Postprocessing/ConcatenatedReadStream.cs
… src/Build/Logging/BinaryLogger/Postprocessing/StreamExtensions.cs
… src/Build/Logging/BinaryLogger/Postprocessing/TransparentReadStream.cs
… src/Build/Microsoft.Build.csproj
… src/Shared/BinaryReaderExtensions.cs