# Improve Binlog Reading by using a Custom Binary Reader 
- Added Buffered-Binary-Reader that combines MemoryStream and BinaryReader. 
- This allows `Read7BitEncodedInt()` to inline `Stream.ReadByte()`, saving 1-5 virtual function calls per int reads.
- Added BulkRead7BitEncodedInt() to further optimize multiple read operations to further save on memory allocations.
- These changes resulted in 2-4X speed up when tested reading a very large binlog.
- Added a IBinaryReader interface to make switching back to BCL BinaryReader easy.
# Improve Binlog Writer's Write7IntEncodedInt().
- Write to a buffer before writing to the stream.  Reduce from 1 - 5 virtual function calls to 1.
- This change reduces the total time in the function, but since it is not on a critical path, its benefit is negligible.